+++ Seminar01/HomeWork/src/HotDrinksSaleMachine.java
@@ -12,7 +12,7 @@ public class HotDrinksSaleMachine implements VendingMachine{
         return null;
     }
 
-    public ExtraHotDrinks getDrink(int volume) {
+    public ExtraHotDrinks getDrink(double volume) {
         for (ExtraHotDrinks drink: hotDrinks) {
             if (drink.getVolume() == volume) {
                 return drink;
@@ -21,7 +21,7 @@ public class HotDrinksSaleMachine implements VendingMachine{
         return null;
     }
 
-    public ExtraHotDrinks getDrink(double temperature) {
+    public ExtraHotDrinks getDrink(int temperature) {
         for (ExtraHotDrinks drink: hotDrinks) {
             if (drink.getTemperature() == temperature) {
                 return drink;
@@ -30,7 +30,7 @@ public class HotDrinksSaleMachine implements VendingMachine{
         return null;
     }
 
-    public ExtraHotDrinks getDrink(String name, int volume, double temperature) {
+    public ExtraHotDrinks getDrink(String name, double volume, int temperature) {
         for (ExtraHotDrinks drink: hotDrinks) {
             if (drink.getName().equals(name) && drink.getVolume() == volume && drink.getTemperature() == temperature) {
                 return drink;
